requirements.txt
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cu118
--extra-index-url https://www.paddlepaddle.org.cn/packages/stable/cu118/

torch
torchvision
torchaudio

ultralytics==8.3.27
matplotlib==3.8.2
pillow==10.1.0
opencv-python==4.10.0.84 
pandas==2.1.3
numpy==1.26.4
scikit-image==0.22.0
geopandas==1.1.1
shapely==2.1.1

paddlepaddle-gpu==3.1.0
paddleocr


